Fixed movment code.
Player now moves right across the scene very slowly

+++ godottesting/Assets/Scripts/Player.cs
@@ -3,9 +3,11 @@ using System;
 
 public partial class Player : Node2D
 {
-	
-	public override void _PhysicsProcess(double delta) {
-		GlobalPosition += new Vector2(40*delta,0);
+
+	public override void _PhysicsProcess(double delta)
+	{
+		GlobalPosition += new Vector2(40 * (float)delta, 0);
+		
 	}
